[수정] 스크롤로 이미지 확대 애니메이션을 커스텀훅으로 분리
